requirements.txt
python-dotenv
openai
kedro~=0.19.11
kedro-datasets>=3.0
kedro-viz>=6.7.0
kedro[jupyter]~=0.19.11
ipython>=8.10
jupyterlab>=3.0
notebook
pandas